explicitly set image dimensions after transform
the Config field was added to gif.GIF in go1.5.  It has a fallback for
go1.4 code where a zero value Config is determined by the Max bound of
the first frame[0].  However, in our case we are transforming each
frame of the gif in place, so the original Config value is there from
when the image was first decoded.  So instead, we now manually set the
Config height and width using the same method as the gif package itself.

[0]: https://github.com/golang/go/commit/6abfdc3fdd49d7691e1cdc3028849fa40e4a41f2