respect the disposal method of each frame
I'm not 100% sure DisposalBackground is being handled properly here. The
gif.GIF type does expose the BackgroundIndex, but not an easy way to
access the global color table that this is an index for.  GIF.Config
seems to be based on the global color table, but provides no direct way
to access it.  I'm assuming the gif package handles this itself then?

Fixes #3